Strings that consist of only the character 0 are not considered empty.
Applications on our site typically will have a string that is just "0".
phpDataMapper would typically munge this to a NULL value and insert that
into the database instead. This is due to the fact that the PHP function
empty() considers the string '0' to be an empty value.

I added a test to public function isEmpty in class phpDataMapper_Base in
the Base.php file. This will allow string values that contain the "0"
character to be considered not empty.

All tests passed.